updates to stock form
NOTE!!! - removing the disabled attribute as this was throwing a 500
error when trying to add or remove stock.

We need to revisit this….hiding an input can easily be hacked on the
front end by using dev tools to unhide it and change the value and add
stock to a different user.